0611プログラムで使用しているライブラリを移動(DXライブラリ、Effekseer)、エフェクト追加、バグ修正

+++ MyGame/Actor/ActorManager.cpp
@@ -63,7 +63,7 @@ ActorPtr ActorManager::find(const std::string & name) const{
 }
 
 unsigned int ActorManager::count() const{
-	return actors_.size();
+	return (unsigned int)actors_.size();
 }
 
 void ActorManager::each(std::function<void(const ActorPtr&)> fn) const
